+++ composeApp/src/commonMain/kotlin/de/medizininformatikinitiative/medgraph/ui/searchengine/QueryViewModel.kt
@@ -3,7 +3,6 @@ package de.medizininformatikinitiative.medgraph.ui.searchengine
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import cafe.adriel.voyager.core.model.ScreenModel
 import de.medizininformatikinitiative.medgraph.searchengine.model.RawQuery
 
 /**
@@ -16,7 +15,9 @@ class QueryViewModel {
     var queryText by mutableStateOf("")
     var productQueryText by mutableStateOf("")
     var substanceQueryText by mutableStateOf("")
+    var dosageQueryText by mutableStateOf("")
+    var doseFormQueryText by mutableStateOf("")
 
-    fun createQuery() = RawQuery(queryText, productQueryText, substanceQueryText)
+    fun createQuery() = RawQuery(queryText, productQueryText, substanceQueryText, dosageQueryText, doseFormQueryText)
 
 }
