+++ medgraph/src/test/java/de/medizininformatikinitiative/medgraph/searchengine/algorithm/querymanagement/DoseFormQueryRefinerImplTest.java
@@ -21,7 +21,7 @@ import static org.junit.jupiter.api.Assertions.*;
 /**
  * @author Markus Budeus
  */
-public class DoseFormQueryRefinerTest extends UnitTest {
+public class DoseFormQueryRefinerImplTest extends UnitTest {
 
 	private DoseFormQueryRefiner sut;
